Fixing deployment issues
