+++ src/FairPlayCombinedSln/FairPlayCombined.Models/AzureMaps/SearchPOICategoryTree/POICategoryModel.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
+﻿#pragma warning disable S101 // Types should be named in PascalCase
 namespace FairPlayCombined.Models.AzureMaps.SearchPOICategoryTree
 {
     public class POICategoryModel
@@ -14,3 +9,4 @@ namespace FairPlayCombined.Models.AzureMaps.SearchPOICategoryTree
         public string[]? Synonyms { get; set; }
     }
 }
+#pragma warning restore S101 // Types should be named in PascalCase
